GIT Project/requirements_streamlit_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=1.3.0=py39haa95532_0
aiohttp=3.8.3=py39h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py39haa95532_0
attrs=22.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py39haa95532_0
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2023.01.10=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
chardet=5.1.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
colorama=0.4.6=py39haa95532_0
cryptography=39.0.1=py39h21b164f_0
decorator=5.1.1=pypi_0
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=pypi_0
ffmpeg=4.2.2=he774522_0
flatbuffers=2.0.0=h6c2663c_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py39h2bbff1b_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py39hc60d5dd_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=3.7.0=py39h3de5c98_0
hdf5=1.10.6=h1756f20_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py39haa95532_0
importlib-metadata=6.0.0=py39haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.2=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=pypi_0
keras=2.10.0=py39haa95532_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
lerc=3.0=hd77b12b_0
libclang=12.0.0=default_h627e005_2
libcurl=7.88.1=h86230a5_0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libssh2=1.10.0=hcd4344a_0
libtiff=4.5.0=h6c2663c_2
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.1=py39haa95532_0
markupsafe=2.1.2=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
multidict=6.0.2=py39h2bbff1b_0
mysql-connector=2.2.9=pypi_0
mysql-connector-python=8.0.32=pypi_0
numpy=1.24.2=pypi_0
numpy-base=1.23.5=py39h4da318b_0
oauthlib=3.2.1=py39haa95532_0
opencv=4.6.0=py39hf11a4ad_3
openssl=1.1.1t=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.0=py39haa95532_0
pandas=1.5.3=pypi_0
pcre=8.45=hd77b12b_0
pillow=9.4.0=pypi_0
pip=23.0.1=py39haa95532_0
pooch=1.4.0=pyhd3eb1b0_0
protobuf=3.20.3=py39hd77b12b_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py39haa95532_0
pyopenssl=23.0.0=py39haa95532_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.16=h6244533_2
python-dateutil=2.8.2=pypi_0
python-flatbuffers=2.0=pyhd3eb1b0_0
pytz=2022.7.1=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.2=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.10.0=py39h321e85e_1
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sqlite=3.41.1=h2bbff1b_0
streamlit=1.20.0=pypi_0
streamlit-lottie=0.0.3=pypi_0
streamlit-option-menu=0.3.2=pypi_0
tensorboard=2.10.0=py39haa95532_0
tensorboard-data-server=0.6.1=py39haa95532_0
tensorboard-plugin-wit=1.8.1=py39haa95532_0
tensorflow=2.10.0=mkl_py39ha510bab_0
tensorflow-base=2.10.0=mkl_py39h6a7f48e_0
tensorflow-estimator=2.10.0=py39haa95532_0
termcolor=2.1.0=py39haa95532_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.4.0=py39haa95532_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.15=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.3.1=pypi_0
werkzeug=2.2.2=py39haa95532_0
wheel=0.38.4=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
wrapt=1.14.1=py39h2bbff1b_0
xz=5.2.10=h8cc25b3_1
yarl=1.8.1=py39h2bbff1b_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
